Working LED with Twitch API
In this version, the script successfully scans if any of the user's followed streamers are live. If one is live, it sends a message to the Arduino that turns the led on in one minute. It seems as though time.sleep() halts both the python script and the arduino, whereas delay() only pauses the arduino, which is my theory on why it takes one minute (which makes sense, time.sleep() would cut the communication for a minute). The user_id header still needs to be configured dynamically. During testing, I used my own directly in the url on line 21. I am considering altering the file structure and content, as not everything is needed (as some files were copied from the Twitch CMD project. Additionally, the main component of the next commit should be changing the output of the light on the arduino, using a circuit board. A schematic should accompany it.